Making admin and employee base

+++ API/Repositories/ClientRepository.cs
@@ -16,17 +16,30 @@ namespace API.Repositories
 
         public List<Client> GetAllClients()
         {
-            return [.. _db.Clients];
+            return [.. _db.Clients.Where(c => c.ClientType == Models.Enums.ClientTypeEnum.Client)];
         }
 
         public Client GetClient(int Id)
         {
-            return _db.Clients.AsNoTracking().FirstOrDefault(c => c.IdClient == Id);
+            return _db.Clients.AsNoTracking().FirstOrDefault(c => c.ClientType == Models.Enums.ClientTypeEnum.Client && c.IdClient == Id);            
         }
         
+        public Client GetCompleteClient(int Id)
+        {
+            var client = _db.Clients.AsNoTracking().FirstOrDefault(c => c.ClientType == Models.Enums.ClientTypeEnum.Client && c.IdClient == Id);            
+
+            //Fill client's schedule from requisition date
+            if(client != null)
+            {
+
+            }
+
+            return client;
+        }
+
         public Client GetClientByEmail(string Email)
         {
-            return _db.Clients.AsNoTracking().FirstOrDefault(c => c.Email == Email);
+            return _db.Clients.AsNoTracking().FirstOrDefault(c => c.ClientType == Models.Enums.ClientTypeEnum.Client && c.Email == Email && c.Active);
         }
 
         public void PostClient(Client client)
@@ -42,9 +55,30 @@ namespace API.Repositories
         }
 
         public void DeleteClient(Client client)
-        {         
-            _db.Clients.Remove(client);
-            _db.SaveChanges();
+        {
+            if (ClientHasSchedule(client.IdClient))
+            {
+                client.Active = false;
+                client.UpdatedAt = DateTime.UtcNow;
+
+                _db.Clients.Update(client);
+                _db.SaveChanges();
+            }
+            else
+            {
+                _db.Clients.Remove(client);
+                _db.SaveChanges();
+            }
+        }
+
+        private bool ClientHasSchedule(int Id)
+        {
+            var client = _db.Clients.FromSqlInterpolated($"SELECT C.IdClient FROM Clients AS C JOIN Schedules AS S ON S.IdClient = C.IdClient WHERE C.IdClient = {Id}").FirstOrDefault();
+
+            if (client == null)
+                return false;
+
+            return true;
         }
     }
 }
